倍精度に戻した・vortexとshearのflgを追加

+++ program/mkdata/modules/constants.f90
@@ -1,4 +1,4 @@
 module constants
  implicit none
- real(4), parameter ::   g = 9.80665 ! m/s^2
+ real(8), parameter ::   g = 9.80665 ! m/s^2
 end module constants
